+++ CHANGELOG.rst
@@ -24,6 +24,9 @@ Bug fixes and small improvements:
 - Change default HTML theme excluded line color. (:issue:`1179`)
 - Move theme colors to a separate CSS file so they can be overridden with a simple
   `--html-template-directory` directory with only the style.colors.css file (:issue:`1179`)
+- Fix regular expression for detecting a version mismatch between ``gcc`` and ``gcov``. (:issue:`1187`)
+- Improve logging messages for GCOV execution errors to not print information twice and add trace
+  messages for gcov execution. (:issue:`1187`)
 
 Documentation:
 
@@ -59,6 +62,7 @@ Known bugs:
 - :ref:`fix_1171_1`
 - :ref:`fix_1171_2`
 - :ref:`fix_1176`
+- :ref:`fix_1187`
 
 Breaking changes: